+++ 42/media/lua/shared/PhobosLib.lua
@@ -5,7 +5,7 @@
 ---------------------------------------------------------------
 
 PhobosLib = PhobosLib or {}
-PhobosLib.VERSION = "1.2.0"
+PhobosLib.VERSION = "1.4.0"
 
 require "PhobosLib_Util"
 require "PhobosLib_Fluid"
@@ -15,3 +15,4 @@ require "PhobosLib_Quality"
 require "PhobosLib_Hazard"
 require "PhobosLib_Skill"
 require "PhobosLib_Reset"
+require "PhobosLib_Validate"